+++ applications/desktop/src/common/commands/by-menu/app.ts
@@ -1,0 +1,19 @@
+export const About = {
+  name: "About",
+  mapToElectronRole: "about",
+};
+
+export const Hide = {
+  name: "Hide",
+  mapToElectronRole: "hide",
+};
+
+export const HideOthers = {
+  name: "HideOthers",
+  mapToElectronRole: "hideOthers",
+};
+
+export const Unhide = {
+  name: "Unhide",
+  mapToElectronRole: "unhide",
+};
