+++ applications/desktop/src/common/commands/by-menu/view.ts
@@ -1,0 +1,29 @@
+export const Reload = {
+  name: "Reload",
+  mapToElectronRole: "forceReload",
+};
+
+export const Fullscreen = {
+  name: "Fullscreen",
+  mapToElectronRole: "togglefullscreen",
+};
+
+export const DevTools = {
+  name: "DevTools",
+  mapToElectronRole: "toggleDevTools",
+};
+
+export const ZoomReset = {
+  name: "ZoomReset",
+  mapToElectronRole: "resetZoom",
+};
+
+export const ZoomIn = {
+  name: "ZoomIn",
+  mapToElectronRole: "zoomIn",
+};
+
+export const ZoomOut = {
+  name: "ZoomOut",
+  mapToElectronRole: "zoomOut",
+};
