+++ applications/desktop/src/common/commands/by-menu/window.ts
@@ -1,0 +1,4 @@
+export const Minimize = {
+  name: "Minimize",
+  mapToElectronRole: "hide",
+};
