+++ applications/desktop/src/common/commands/index.ts
@@ -1,4 +1,7 @@
-export * from "./contents";
-export * from "./export-pdf";
-export * from "./publish-gist";
-export * from "./roles";
+export * from "./by-menu/app";
+export * from "./by-menu/cell";
+export * from "./by-menu/edit";
+export * from "./by-menu/file";
+export * from "./by-menu/runtime";
+export * from "./by-menu/view";
+export * from "./by-menu/window";
