--- applications/desktop/src/common/commands/roles.ts
@@ -1,3 +1,0 @@
-export const Copy = {
-  role: "copy",
-};
