+++ applications/desktop/src/common/commands/utils/auth.ts
