+++ applications/desktop/src/common/commands/utils/directories.ts
@@ -3,7 +3,7 @@ import { ContentRef } from "@nteract/types";
 import { remote } from "electron";
 import * as fs from "fs";
 import * as path from "path";
-import { DesktopStore } from "../notebook/store";
+import { DesktopStore } from "../../../notebook/store";
 
 export const currentDocumentDirectory =
   (store: DesktopStore, contentRef: ContentRef) =>
